% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_station_metadata.R
\name{all_station_metadata}
\alias{all_station_metadata}
\title{Get Processed Station Metadata}
\usage{
all_station_metadata(
  country = c("zm", "mw"),
  format = c("wide", "long", "nested")
)
}
\arguments{
\item{country}{A character vector specifying the country code for which station metadata should be retrieved and processed.}

\item{format}{A character vector indicating the desired format of the processed data. It can be "wide", "long", or "nested".}
}
\value{
Depending on the specified format, the function returns the processed station metadata in either wide, long, or nested format.
}
\description{
This function retrieves and processes station metadata for the specified country and format. The station metadata includes information about station IDs and their associated definitions.
}
\examples{
# Retrieve and process station metadata for country "zm" in wide format
#all_station_metadata(country = "zm", format = "wide")

# Retrieve and process station metadata for countries "zm" and "mw" in long format
#all_station_metadata(country = c("zm", "mw"), format = "long")

# Retrieve and process station metadata for country "zm" in nested format
#all_station_metadata(country = "zm", format = "nested")
}
