% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_annual_summaries_definitions.R
\name{build_annual_summaries_definitions}
\alias{build_annual_summaries_definitions}
\title{Get annual summaries definitions}
\usage{
build_annual_summaries_definitions(
  data_name,
  by_definitions_list,
  data_definitions_list
)
}
\arguments{
\item{data_name}{The name of the data.}

\item{by_definitions_list}{A list containing definitions for start of rains, end of rains, end of season, and seasonal length.}

\item{data_definitions_list}{A list containing data definitions.}
}
\value{
A JSON representation of annual summaries definitions.
}
\description{
Retrieves annual summaries definitions including start of rains, end of rains, end of season, seasonal length,
and annual rainfall summaries.
}
\examples{
# Example usage:
#get_annual_summaries_definitions("data_name", by_definitions_list, data_definitions_list)
}
