% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate_definitions_data.R
\name{collate_definitions_data}
\alias{collate_definitions_data}
\title{Collate Definitions Data for Climatic Analysis from R-Instat}
\usage{
collate_definitions_data(
  main_data_set = "ghana",
  data_by_year = "ghana_by_station_year",
  data_by_year_month = NULL,
  crop_data = "crop_def",
  rain = data_book$get_climatic_column_name(data_name = "ghana", col_name = "rain"),
  tmin = data_book$get_climatic_column_name(data_name = "ghana", "temp_min"),
  tmax = data_book$get_climatic_column_name("ghana", "temp_max"),
  year = data_book$get_climatic_column_name("ghana", "year"),
  month = data_book$get_climatic_column_name("ghana", "month")
)
}
\arguments{
\item{main_data_set}{The name of the main data set, default is "ghana".}

\item{data_by_year}{The name of the data set that contains data aggregated by year, default is "ghana_by_station_year".}

\item{data_by_year_month}{The name of the data set that contains data aggregated by year and month, default is NULL.}

\item{crop_data}{The name of the crop data set, default is "crop_def".}

\item{rain}{The name of the column containing rainfall data.}

\item{tmin}{The name of the column containing minimum temperature data.}

\item{tmax}{The name of the column containing maximum temperature data.}

\item{year}{The name of the column containing year data.}

\item{month}{The name of the column containing month data.}
}
\value{
A JSON-formatted string that contains the aggregated data definitions.
}
\description{
This function aggregates various climatic data definitions, including annual summaries,
temperature summaries, crop data, and probabilities of season starts. It is designed to work
within a specific context that involves climatic data processing and analysis, particularly
focusing on data related to Ghana's climate. The function uses multiple sources of data
and calculations to generate a comprehensive JSON-formatted summary.
}
\examples{
#data_book <- list(get_climatic_column_name = function(data_name, col_name) { return(col_name) },
#                  get_calculations = function(data_name) { list() },
#                  get_data_frame_metadata = function(data_name) { list() })
#collate_definitions_data(data_book = data_book)

}
