% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_definitions.R
\name{data_definitions}
\alias{data_definitions}
\title{Get Data Column Names}
\usage{
data_definitions(station_id)
}
\arguments{
\item{station_id}{The ID of the station.}
}
\value{
A list of data definitions based on the station ID.
}
\description{
Get Data Column Names
}
\examples{
# todo
}
