% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bucket_name.R
\name{get_bucket_name}
\alias{get_bucket_name}
\title{Get Bucket Name}
\usage{
get_bucket_name(country = c("mw", "zm"))
}
\arguments{
\item{country}{A character vector specifying the country or countries from which to update the data. Options are \code{"mz"} and \code{"zm"}.}
}
\value{
Returns the name of the bucket for the Malawi or Zambia data.
}
\description{
Get Bucket Name
}
\examples{
get_bucket_name("mw")
}
