% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_daily_data.R
\name{get_daily_data}
\alias{get_daily_data}
\title{Get Daily Data}
\usage{
get_daily_data(country = c("mw", "zm", "zm_test"), station_id)
}
\arguments{
\item{country}{A character vector specifying the country or countries from which to get the data. Options are \code{"mz"} and \code{"zm"}.}

\item{station_id}{A character string specifying the ID of the station for which to get the daily data.}
}
\value{
A data frame containing the daily data for the specified station and country.
}
\description{
Get Daily Data
}
\examples{
#
}
