% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_definitions_data.R
\name{get_definitions_data}
\alias{get_definitions_data}
\title{Get Daily Data}
\usage{
get_definitions_data(country = c("mw", "zm"), station_id)
}
\arguments{
\item{country}{A character vector specifying the country or countries from which to get the definitions data. Options are \code{"mz"} and \code{"zm"}.}

\item{station_id}{A character string specifying the ID of the station for which to get the definitions data.}
}
\value{
A data frame containing daily data based on the station ID.
}
\description{
Get Daily Data
}
\examples{
#
}
