% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_temperature_summary_definitions.R
\name{get_temperature_summary_definitions}
\alias{get_temperature_summary_definitions}
\title{Get temperature summary definitions}
\usage{
get_temperature_summary_definitions(
  tmin = data_book$get_climatic_column_name(data_name, "temp_min"),
  tmax = data_book$get_climatic_column_name(data_name, "temp_max"),
  year = data_book$get_climatic_column_name(data_name, "year"),
  month = data_book$get_climatic_column_name(data_name, "month"),
  by_definition_list = data_defs,
  by_definition_2_list = NULL
)
}
\arguments{
\item{tmin}{Character vector specifying the minimum temperature.}

\item{tmax}{Character vector specifying the maximum temperature.}

\item{year}{Character vector specifying the year.}

\item{month}{Character vector specifying the month.}

\item{by_definition_list}{A list of temperature summaries by definition.}

\item{by_definition_2_list}{An optional second list of temperature summaries by definition.}
}
\value{
A list containing temperature summary definitions.
}
\description{
Retrieves temperature summary definitions based on provided parameters.
}
\examples{
# Example usage:
#get_temperature_summary_definitions(by_definition_list = my_definition_list, by_definition_2_list = my_definition_list_2)
}
