% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annual_rain_definitions.R
\name{get_total_rain_counts}
\alias{get_total_rain_counts}
\title{Get annual rain definitions}
\usage{
get_total_rain_counts(
  data_name,
  data_by_year,
  rain_name = data_book$get_climatic_column_name(data_name = "ghana", col_name = "rain")
)
}
\arguments{
\item{data_name}{The name of the main data frame.}

\item{data_by_year}{The definitions file for the data_by_year data frame.}

\item{rain_name}{The rain variable in the data frame.}
}
\value{
A list representation of annual rain definitions.
}
\description{
Retrieves annual rain definitions.
}
\examples{
# Example usage:
#get_total_rain_counts(annual_rain, ghana_defs)

}
