% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_metadata.R
\name{station_metadata}
\alias{station_metadata}
\title{Get Station Metadata}
\usage{
station_metadata(country = c("mw", "zm"))
}
\arguments{
\item{country}{A character vector specifying the country or countries from which to get the metadata. Options are \code{"mz"} and \code{"zm"}.}
}
\value{
This function gets the metadata for the specified station in the specified country.
}
\description{
Get Station Metadata
}
\examples{
#todo
}
