% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_total_temperature_summaries.R
\name{build_total_temperature_summaries}
\alias{build_total_temperature_summaries}
\title{Calculate total temperature summaries}
\usage{
build_total_temperature_summaries(
  tmin = data_book$get_climatic_column_name(data_name, "temp_min"),
  tmax = data_book$get_climatic_column_name(data_name, "temp_max"),
  year = data_book$get_climatic_column_name(data_name, "year"),
  month = data_book$get_climatic_column_name(data_name, "month"),
  by_definition_list = data_defs,
  by_definition_2_list = by_definition_2_list
)
}
\arguments{
\item{tmin}{Character vector specifying the minimum temperature.}

\item{tmax}{Character vector specifying the maximum temperature.}

\item{year}{Character vector specifying the year.}

\item{month}{Character vector specifying the month.}

\item{by_definition_list}{A list of temperature summaries by definition.}

\item{by_definition_2_list}{An optional second list of temperature summaries by definition.}
}
\value{
A JSON-formatted string containing total temperature summaries.
}
\description{
Calculates total temperature summaries based on provided parameters.
}
\examples{
# Example usage:
#total_temperature_summaries(tmin = "tmin", tmax = "tmax", year = "year", month = "month", 
#                             by_definition_list = my_definition_list, by_definition_2_list = my_definition_list_2)
}
