% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annual_rain_definitions.R
\name{get_annual_rain_definitions}
\alias{get_annual_rain_definitions}
\title{Get annual rain definitions}
\usage{
get_annual_rain_definitions(sum_rain = NULL, n_rain_def = NULL, data)
}
\arguments{
\item{annual_rain}{The annual rain data.}

\item{ghana_defs}{The Ghana definitions.}
}
\value{
A list representation of annual rain definitions.
}
\description{
Retrieves annual rain definitions.
}
\examples{
# Example usage:
#get_annual_rain_definitions(annual_rain, ghana_defs)

# In Progress.
}
