% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_daily_data.R
\name{get_daily_data}
\alias{get_daily_data}
\title{Get Daily Data}
\usage{
get_daily_data(country = c("mw", "zm"), station_id)
}
\arguments{
\item{country}{}

\item{station_id}{}
}
\description{
Get Daily Data
}
