% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_definitions_data.R
\name{get_definitions_data}
\alias{get_definitions_data}
\title{Get Daily Data}
\usage{
get_definitions_data(country = c("mw", "zm"), station_id)
}
\arguments{
\item{country}{}

\item{station_id}{}
}
\description{
Get Daily Data
}
